byte Max is now int Highest!

+++ Xamarin.ExposureNotification/ExposureNotification.android.cs
@@ -206,7 +206,7 @@ namespace Xamarin.ExposureNotifications
 			return new ExposureDetectionSummary(
 				summary.DaysSinceLastExposure,
 				(ulong)summary.MatchedKeyCount,
-				(byte)summary.MaximumRiskScore,
+				summary.MaximumRiskScore,
 				summary.GetAttenuationDurationsInMinutes()
 					.Select(a => TimeSpan.FromMinutes(a)).ToArray(),
 				summary.SummationRiskScore);
